@@ -1,8 +1,11 @@
 module github.com/yl2chen/cidranger
 
-go 1.13
+go 1.18
+
+require github.com/stretchr/testify v1.6.1
 
 require (
-	github.com/stretchr/testify v1.6.1
-	gopkg.in/yaml.v2 v2.2.2 // indirect
+	github.com/davecgh/go-spew v1.1.0 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
+	gopkg.in/yaml.v3 v3.0.0-20200313102051-9f266ea9e77c // indirect
 )
